Revert "Driver test unit, big pull. Look through carefully! Was a bit messy."
