+++ pyscan/__init__.py
@@ -2,4 +2,3 @@ from pyscan.measurement import *
 from pyscan.drivers import *
 from pyscan.general import *
 from pyscan.plotting import *
-from pyscan.drivers.testing import *
