+++ pyscan/drivers/__init__.py
@@ -1,7 +1,6 @@
 # Objects
 from .instrument_driver import InstrumentDriver
 
-
 # Instrument Drivers
 from .agilent33500 import Agilent33500
 from .agilent34410 import Agilent34410
@@ -75,11 +74,8 @@ try:
 except ModuleNotFoundError:
     print('Thorlabs Kinesis not found, ThorlabsMFF101 not loaded')
 
-
 # Methods
 from .new_instrument import new_instrument
 
-
 # Test Devices
-from .testing.test_voltage import TestVoltage
-from .testing.test_instrument_driver import TestInstrumentDriver
+from .test_voltage import TestVoltage
